+++ src/hooks/index.ts
@@ -1,4 +1,4 @@
 export * from './useVerifyUser';
 export * from './useLogout';
 export * from './useLike';
-export * from './useInfinityScroll';
+export * from './useIntersect';
